% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFossilSeriesAbundanceTable.R
\name{checkFossilSeriesAbundanceTable}
\alias{checkFossilSeriesAbundanceTable}
\title{Check Fossil Series Abundance Table}
\usage{
checkFossilSeriesAbundanceTable(
  fossilSeries,
  minSampleSize,
  maxSampleSize,
  maxDominance
)
}
\arguments{
\item{fossilSeries}{The}

\item{minSampleSize}{The}

\item{maxSampleSize}{The}

\item{maxDominance}{The}
}
\value{
The
}
\description{
The
}
\details{
The
}
