% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleProperties.R
\name{getSampleProperties}
\alias{getSampleProperties}
\title{Get Sample Properties}
\usage{
getSampleProperties(
  simTimeVar,
  fossilSeries,
  eventStartEndTimes,
  initialBackgroundIntervalIncluded,
  backgroundStartEnd
)
}
\arguments{
\item{simTimeVar}{The}

\item{fossilSeries}{The}

\item{eventStartEndTimes}{The}

\item{initialBackgroundIntervalIncluded}{The}

\item{backgroundStartEnd}{The}
}
\value{
The
}
\description{
The
}
\details{
The
}
