% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGradientKDE.R
\name{plotGradientKDE}
\alias{plotGradientKDE}
\title{Plot Kernel Density Estimates of Species Abundance Across a Focal Gradient}
\usage{
plotGradientKDE(
  speciesKDEs,
  fullGradientRange,
  xlim = NULL,
  ylim = c(0, 1),
  logY = FALSE
)
}
\arguments{
\item{speciesKDEs}{}

\item{fullGradientRange}{}

\item{xlim}{}

\item{ylim}{}

\item{logY}{}
}
\value{
The
}
\description{
The
}
\details{
The
}
