% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmapComparison.R
\name{plotHeatmapComparison}
\alias{plotHeatmapComparison}
\title{Plots a Heatmap Comparison}
\usage{
plotHeatmapComparison(
  x,
  y,
  z,
  xlim = range(x, finite = TRUE),
  ylim = range(y, finite = TRUE),
  zlim = range(z, finite = TRUE),
  xlog = FALSE,
  ylog = FALSE,
  xtick = pretty(x),
  ytick = pretty(y),
  contourLevels = NULL,
  nlevels = 10,
  contourLineLevels = NULL,
  contour.lwd = 2,
  additionalFeatures = NULL,
  key.axes = NULL,
  palette = "plasma",
  xlab = "x",
  ylab = "y",
  main = "main title",
  gradientKeyLabel = "color gradient key",
  mtext_line = 3,
  margins = c(5, 6, 5, 5)
)
}
\arguments{
\item{x, y}{The}

\item{z}{The}

\item{xlim, ylim, zlim}{The}

\item{xlog, ylog}{The}

\item{xtick, ytick}{The}

\item{contourLevels}{The}

\item{nlevels}{The}

\item{contourLineLevels}{The}

\item{contour.lwd}{The}

\item{additionalFeatures}{The}

\item{key.axes}{The}

\item{palette}{The}

\item{xlab, ylab}{The}

\item{main}{The}

\item{gradientKeyLabel}{The}

\item{mtext_line}{The}

\item{margins}{The}
}
\value{
The
}
\description{
The
}
\details{
The
}
