% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyCommunityEcology.R
\name{quantifyCommunityEcology}
\alias{quantifyCommunityEcology}
\title{Quantify Community Ecology From a Time Series of Simulated Fossil Assemblages}
\usage{
quantifyCommunityEcology(
  origAbundData,
  fossilSeries,
  useTransformedRelAbundance = TRUE,
  projectIntoOrigDCA = TRUE,
  powerRootTransform = 1
)
}
\arguments{
\item{origAbundData}{The original matrix of abundance data, to be used to
project the simulated data into the same detrended correspondence analysis (DCA) space.}

\item{fossilSeries}{The}

\item{useTransformedRelAbundance}{Should the DCA be analyzed}

\item{projectIntoOrigDCA}{Should the new simulated data be projected in the DCA generated by analyzing the original data? This is \code{TRUE} by default, which is what most users will likely use, as it is the preferable way to consider how the new simulated data relates to the original data.}

\item{powerRootTransform}{The power-root transform to be used on the abundance data before applying the DCA. By default this is 1, which means the data is not transformed at all. Note that the power-root transform is only performed if \code{useTransformedRelAbundance = TRUE}.}
}
\value{
A list containing the input data and arguments, as well as the 
DCA score values found
}
\description{
Given the output from functions such as \code{simulateFossilAssemblageSeries}, 
this function calculates various ecological measures that might be used as
'recovered' gradient values to compare against the original generating gradient values.
}
\details{
Quantifies
}
