% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFossilAssemblageSeries.R
\name{sampleFossilAssemblageSeries}
\alias{sampleFossilAssemblageSeries}
\title{Sample Fossil Assemblage Series}
\usage{
sampleFossilAssemblageSeries(
  bioturbIntensity,
  bioturbZoneDepth,
  distBetweenSamples,
  sampleWidth,
  simTimeVar,
  timestepAbundances,
  nSpecimens
)
}
\arguments{
\item{bioturbIntensity}{}

\item{bioturbZoneDepth}{}

\item{distBetweenSamples}{}

\item{sampleWidth}{}

\item{simTimeVar}{}

\item{timestepAbundances}{}

\item{nSpecimens}{}
}
\description{
Sample Fossil Assemblage Series
}
\details{
Importantly, this function is where bioturbation is handled
}
