% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupSimulatedGradientChange.R
\name{setupSimulatedGradientChange}
\alias{setupSimulatedGradientChange}
\title{Create a Stochastic Time-Series of Gradient Change For Use in Simulating Assemblage Change}
\usage{
setupSimulatedGradientChange(
  nEvents,
  fullGradientRange,
  peakGradientValue,
  bgGradientValue,
  bgDurationRange,
  transitionDuration,
  eventDuration,
  halfGradientOnly = "full",
  includeInitialBackgroundPhase = TRUE,
  plot = FALSE
)
}
\arguments{
\item{nEvents}{}

\item{fullGradientRange}{}

\item{peakGradientValue}{}

\item{bgGradientValue}{}

\item{bgDurationRange}{}

\item{transitionDuration}{}

\item{eventDuration}{}

\item{halfGradientOnly}{}

\item{includeInitialBackgroundPhase}{}

\item{plot}{}
}
\description{
Given a series of inputs, simulates a sequence
}
