% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSimulationDCA.R
\name{summarizeSimulationDCA}
\alias{summarizeSimulationDCA}
\title{Summarize Simulation DCA}
\usage{
summarizeSimulationDCA(
  sampleData,
  nEvents,
  backgroundUpperBound,
  eventMagnitudesAbsolute,
  eventLowerBound,
  eventUpperBound
)
}
\arguments{
\item{sampleData}{The}

\item{nEvents}{The}

\item{backgroundUpperBound}{The}

\item{eventMagnitudesAbsolute}{The}

\item{eventLowerBound}{The}

\item{eventUpperBound}{The}
}
\value{
The
}
\description{
The
}
\details{
The
}
