% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateImplicitParameters.R
\name{calculateImplicitParameters}
\alias{calculateImplicitParameters}
\title{Calculate Implicit Parameters}
\usage{
calculateImplicitParameters(
  eventChangeScale,
  bgGradientValue,
  fullGradientRange,
  eventSampleWidthRatio = NULL,
  sampleWidth = NULL,
  eventDuration = NULL,
  sedRatePerTimestep = NULL,
  maxSampleTimeStep = 500,
  minSampleTimeStep = 3,
  samplingCompleteness,
  transitionDurationRatio,
  bioturbDepthRatio
)
}
\arguments{
\item{eventChangeScale}{A value indicating the amount relative to the background value (\code{bgGradientValue}) and the maximum possible change as indicated by \code{fullGradientRange} (in other words, simulated change must be within observed gradient, so \code{eventChangeScale} is a proportional multiplier of the total possible change).}

\item{bgGradientValue}{The gradient value expected during background intervals during which no notable excursion is occurring on that environmental gradient.}

\item{fullGradientRange}{A vector of two values giving the minumum and maximum gradient values observed in the empirical data.}

\item{eventSampleWidthRatio}{How long should an event be relative to the amount of time (or sediment) captured within a sedimentary sample? This parameter is used for simulating event duration, sample width and sedimentation rate where any two of these three are defined and the third is not defined.}

\item{sampleWidth}{The 'width' of a sample relative to core depth or outcrop height, usually given in linear units (usually centimeters). For taking sediment samples from a core, this is straightforward (how thick is each sediment sample taken?) but for outcrops this may be more difficult to determine (what is the thickness of a horizon in a shale unit?).}

\item{eventDuration}{The duration (in time-units) of a simulated event during which the environmental gradient is at an excursion 'peak' level.}

\item{sedRatePerTimestep}{The}

\item{maxSampleTimeStep}{The}

\item{minSampleTimeStep}{The}

\item{samplingCompleteness}{The}

\item{transitionDurationRatio}{The}

\item{bioturbDepthRatio}{The}
}
\value{
Returns a list giving the full set of parameters necessary for running \code{\link{simulateFossilAssemblageSeries}}.
}
\description{
Given a set of parameters for simulating fossil assemblages and sampling them as a model time-series, 
 using a particular while some parameterizations
}
\details{
Under the models considered in \code{paleoAM}, some parameterizations may be equivalent, even though the a particular analysis might be better simulated using a particular set of parameters. Allowing various different parameterizations is a useful generalization, but requires translating those equivalent parameters from one set to another (eg specifying parameters A, C & D, but running a simulation that requires parameters A, B & C).

This function mainly exists to calculate unspecified parameters for \code{\link{simulateFossilAssemblageSeries}} and also to identify conflicting parameter specifications.
}
\seealso{
\code{\link{simulateFossilAssemblageSeries}}
}
