% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecoveredTransitionDuration.R
\name{getRecoveredTransitionDuration}
\alias{getRecoveredTransitionDuration}
\title{Measure the Duration of a Transition Period from Recovered Sequence of Fossil Assemblages with DCA-1 Scores}
\usage{
getRecoveredTransitionDuration(
  simRecord,
  bgUpperEnvelope,
  eventLowerEnvelope = NULL,
  returnAsAge = FALSE,
  trueEventDuration = NA,
  plot = FALSE
)
}
\arguments{
\item{simRecord}{}

\item{bgUpperEnvelope}{}

\item{eventLowerEnvelope}{}

\item{returnAsAge}{}

\item{trueEventDuration}{}

\item{plot}{}
}
\description{
Measure the Duration of a Transition Period from Recovered Sequence of Fossil Assemblages with DCA-1 Scores
}
