% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleProperties.R
\name{getSampleProperties}
\alias{getSampleProperties}
\title{Get Sample Properties}
\usage{
getSampleProperties(
  simTimeVar,
  fossilSeries,
  eventStartEndTimes,
  initialBackgroundIntervalIncluded,
  backgroundStartEnd
)
}
\arguments{
\item{simTimeVar}{}

\item{fossilSeries}{}

\item{eventStartEndTimes}{}

\item{initialBackgroundIntervalIncluded}{}

\item{backgroundStartEnd}{}
}
\value{
The
}
\description{
The
}
\details{
The
}
