% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyCommunityEcology.R
\name{quantifyCommunityEcology}
\alias{quantifyCommunityEcology}
\title{Quantify Community Ecology From a Time Series of Simulated Fossil Assemblages}
\usage{
quantifyCommunityEcology(
  origAbundData,
  fossilSeries,
  useTransformedRelAbundance = TRUE,
  projectIntoOrigDCA = TRUE,
  powerRootTransform = 1
)
}
\arguments{
\item{fossilSeries}{The}
}
\value{
A list containing the input data and arguments, as well as the 
DCA score values found
}
\description{
Given the output from functions such as \code{simulateFossilAssemblageSeries}, 
this function calculates various ecological measures that might be used as
'recovered' gradient values to compare against the original generating gradient values.
}
\details{
Quantifies
}
