% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyCommunityEcology.R
\name{quantifyCommunityEcology}
\alias{quantifyCommunityEcology}
\title{Quantify Community Ecology From a Time Series of Simulated Fossil Assemblages}
\usage{
quantifyCommunityEcology(
  origAbundData,
  fossilSeries,
  useTransformedRelAbundance = TRUE,
  projectIntoOrigDCA = TRUE,
  powerRootTransform = 1,
  singularDCA = TRUE,
  inclusiveDCA = FALSE,
  rawDCA = FALSE
)
}
\arguments{
\item{origAbundData}{The original matrix of abundance data, to be used to project the simulated data into the same detrended correspondence analysis (DCA) space.}

\item{fossilSeries}{}

\item{useTransformedRelAbundance}{}

\item{projectIntoOrigDCA}{}

\item{powerRootTransform}{}

\item{singularDCA}{}

\item{inclusiveDCA}{}

\item{rawDCA}{}
}
\description{
Given the output from functions such as \code{simulateFossilAssemblageSeries}, this function calculates various ecological measures that might be used as 'recovered' gradient values to compare against the original generating gradient values.
}
