% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFossilSeries.R
\name{sampleFossilSeries}
\alias{sampleFossilSeries}
\title{Sample Fossil Assemblage Series}
\usage{
sampleFossilSeries(
  bioturbIntensity,
  bioturbZoneDepth,
  distBetweenSamples,
  sampleWidth,
  simTimeVar,
  timestepAbundances,
  nSpecimens
)
}
\arguments{
\item{bioturbIntensity}{The degree of mixing within the bioturbation zone, as a value fbetween 0 and 1. When intensity is 1, a given sample will consist only}

\item{bioturbZoneDepth}{}

\item{distBetweenSamples}{}

\item{sampleWidth}{}

\item{simTimeVar}{}

\item{timestepAbundances}{}

\item{nSpecimens}{The number of specimens selected in each individual sample.}
}
\value{
A list composed of four components:
\code{simTimeVar}, a 
\code{abundanceTable}, 
\code{sampleIntervals},
and \code{bioturbIntervals},
}
\description{
Sample Fossil Assemblage Series
}
\details{
Importantly, this function is where bioturbation is handled
}
