% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMixedAssemblageSample.R
\name{simMixedAssemblageSample}
\alias{simMixedAssemblageSample}
\title{Simulate a Mixed Fossil Assemblage Composed of Communities at Different Gradient Values, and Sample the Lumped Assemblage}
\usage{
simMixedAssemblageSample(
  kdeRescaled,
  probSpeciesOccur,
  gradientValues,
  specimensPerTimestep,
  nSpecimens
)
}
\description{
This function simulate a mixed fossil assemblage by simulating a series of communities across a defined range of gradient values, lumps them into a single mixed assemblage, and then samples that assemblage as defined by the user.
}
