% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFossilAssemblageSeries.R
\name{simulateFossilAssemblageSeries}
\alias{simulateFossilAssemblageSeries}
\title{Simulate Time-Series of Successive Fossil Assemblages}
\usage{
simulateFossilAssemblageSeries(
  kdeRescaled,
  probSpeciesOccur,
  origAbundData,
  eventChangeScale,
  bgGradientValue,
  fullGradientRange,
  eventSampleWidthRatio,
  sampleWidth,
  eventDuration,
  sedRatePerTimestep,
  samplingCompleteness,
  transitionDurationRatio,
  bioturbDepthRatio,
  bioturbIntensity,
  nEvents,
  nSpecimens,
  specimensPerTimestep = 10000,
  halfGradientOnly = "full",
  useTransformedRelAbundance = TRUE,
  projectIntoOrigDCA = TRUE,
  powerRootTransform = 1,
  maxSampleTimeStep = 500,
  minSampleTimeStep = 3,
  singularDCA = TRUE,
  inclusiveDCA = FALSE,
  rawDCA = FALSE,
  includeInitialBackgroundPhase = FALSE,
  plot = FALSE,
  thinOutput = FALSE
)
}
\arguments{
\item{kdeRescaled}{}

\item{probSpeciesOccur}{}

\item{origAbundData}{}

\item{eventChangeScale}{}

\item{eventSampleWidthRatio}{}

\item{sedRatePerTimestep}{}

\item{samplingCompleteness}{}

\item{transitionDurationRatio}{}

\item{bioturbDepthRatio}{}

\item{bioturbIntensity}{}

\item{nSpecimens}{The number of specimens}

\item{specimensPerTimestep}{The number of specimens returned in a given time-step by \code{getTimestepAbundances}, usually set to an unrealistically high number to represent the true 'unsampled' fossil assemblage. Default is 10000.}

\item{useTransformedRelAbundance}{}

\item{projectIntoOrigDCA}{}

\item{powerRootTransform}{}

\item{singularDCA}{}

\item{inclusiveDCA}{}

\item{plot}{}

\item{thinOutput}{Should the output be thinned to just the sample properties and intrinsic variables? Default is FALSE.}
}
\description{
Given a set of parameters and models describing species abundance,
stochastically models changes in an underlying biotic gradient and simulates
ecological change and a sequence of samples representing change in recovered
 fossil assemblages over that interval.
}
