% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFossilAssemblageSeries.R
\name{simulateFossilAssemblageSeries}
\alias{simulateFossilAssemblageSeries}
\title{Simulate Time-Series of Successive Fossil Assemblages}
\usage{
simulateFossilAssemblageSeries(
  kdeRescaled,
  probSpeciesOccur,
  origAbundData,
  eventChangeScale,
  bgGradientValue,
  fullGradientRange,
  eventSampleWidthRatio,
  sampleWidth,
  eventDuration,
  sedRatePerTimestep,
  samplingCompleteness,
  transitionDurationRatio,
  bioturbDepthRatio,
  bioturbIntensity,
  nEvents,
  nSpecimens,
  specimensPerTimestep = 10000,
  halfGradientOnly = "full",
  useTransformedRelAbundance = TRUE,
  projectIntoOrigDCA = TRUE,
  powerRootTransform = 1,
  maxSampleTimeStep = 500,
  minSampleTimeStep = 3,
  singularDCA = TRUE,
  inclusiveDCA = FALSE,
  rawDCA = FALSE,
  includeInitialBackgroundPhase = FALSE,
  plot = FALSE,
  thinOutput = FALSE
)
}
\arguments{
\item{kdeRescaled}{}

\item{probSpeciesOccur}{}

\item{origAbundData}{}

\item{eventChangeScale}{}

\item{bgGradientValue}{The gradient value expected during background intervals during which no notable excursion is occurring on that environmental gradient.}

\item{fullGradientRange}{A vector of two values giving the minumum and maximum gradient values observed in the empirical data.}

\item{eventSampleWidthRatio}{}

\item{sampleWidth}{The 'width' of a sample relative to core depth or outcrop height, usually given in linear units (usually centimeters). For taking sediment samples from a core, this is straightforward (how thick is each sediment sample taken?) but for outcrops this may be more difficult to determine (what is the thickness of a horizon in a shale unit?).}

\item{eventDuration}{The duration (in time-units) of a simulated event during which the environmental gradient is at an excursion 'peak' level.}

\item{sedRatePerTimestep}{}

\item{samplingCompleteness}{}

\item{transitionDurationRatio}{}

\item{bioturbDepthRatio}{}

\item{bioturbIntensity}{}

\item{nEvents}{Number of events to occur in a simulated sequence of gradient change.}

\item{nSpecimens}{The number of specimens}

\item{specimensPerTimestep}{The number of specimens returned in a given time-step by \code{getTimestepAbundances}, usually set to an unrealistically high number to represent the true 'unsampled' fossil assemblage. Default is 10000.}

\item{halfGradientOnly}{Whether to simulate only half of a background-event sequence, either beginning or terminating the simulation at the peak value. Only a single event can be simulated, so \code{nEvents} must be 1.}

\item{useTransformedRelAbundance}{}

\item{projectIntoOrigDCA}{}

\item{powerRootTransform}{}

\item{maxSampleTimeStep}{}

\item{minSampleTimeStep}{}

\item{singularDCA}{}

\item{inclusiveDCA}{}

\item{includeInitialBackgroundPhase}{A logical indicating whether to include a lengthy background phase, for use in calibrating a simulation. This function is mainly for diagnostic purposes and may be removed in future updates.}

\item{plot}{}

\item{thinOutput}{Should the output be thinned to just the sample properties and intrinsic variables? Default is FALSE.}
}
\description{
Given a set of parameters and models describing species abundance,
stochastically models changes in an underlying biotic gradient and simulates
ecological change and a sequence of samples representing change in recovered
 fossil assemblages over that interval.
}
