% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFossilAssemblageSeries.R
\name{simulateFossilAssemblageSeries}
\alias{simulateFossilAssemblageSeries}
\title{Simulate Time-Series of Successive Fossil Assemblages}
\usage{
simulateFossilAssemblageSeries(
  kdeRescaled,
  probSpeciesOccur,
  origAbundData,
  eventChangeScale,
  bgGradientValue,
  fullGradientRange,
  eventSampleWidthRatio,
  sampleWidth,
  eventDuration,
  sedRatePerTimestep,
  samplingCompleteness,
  transitionDurationRatio,
  bioturbDepthRatio,
  bioturbIntensity,
  nEvents,
  nSpecimens,
  specimensPerTimestep = 10000,
  halfGradientOnly = "full",
  useTransformedRelAbundance = TRUE,
  projectIntoOrigDCA = TRUE,
  powerRootTransform = 1,
  maxSampleTimeStep = 500,
  minSampleTimeStep = 3,
  singularDCA = TRUE,
  inclusiveDCA = FALSE,
  rawDCA = FALSE,
  includeInitialBackgroundPhase = FALSE,
  plot = FALSE,
  thinOutput = FALSE
)
}
\arguments{
\item{kdeRescaled}{The list of modeled KDEs for species abundance, output from \code{\link{getSpeciesSpecificRescaledKDE}}.}

\item{probSpeciesOccur}{The output from \code{\link{getProbOccViaPresAbs}}}

\item{eventChangeScale}{A value indicating the amount relative to the background value (\code{bgGradientValue}) and the maximum possible change as indicated by \code{fullGradientRange} (in other words, simulated change must be within observed gradient, so \code{eventChangeScale} is a proportional multiplier of the total possible change).}

\item{bgGradientValue}{The gradient value expected during background intervals during which no notable excursion is occurring on that environmental gradient.}

\item{fullGradientRange}{A vector of two values giving the minumum and maximum gradient values observed in the empirical data.}

\item{eventSampleWidthRatio}{How long should an event be relative to the amount of time (or sediment) captured within a sedimentary sample? This parameter is used for simulating event duration, sample width and sedimentation rate where any two of these three are defined and the third is not defined.}

\item{sampleWidth}{The 'width' of a sample relative to core depth or outcrop height, usually given in linear units (usually centimeters). For taking sediment samples from a core, this is straightforward (how thick is each sediment sample taken?) but for outcrops this may be more difficult to determine (what is the thickness of a horizon in a shale unit?).}

\item{eventDuration}{The duration (in time-units) of a simulated event during which the environmental gradient is at an excursion 'peak' level.}

\item{sedRatePerTimestep}{The}

\item{samplingCompleteness}{The}

\item{transitionDurationRatio}{The}

\item{bioturbDepthRatio}{The}

\item{bioturbIntensity}{The degree of mixing within the bioturbation zone, as a value fbetween 0 and 1. When intensity is 1, a given sample will consist only}

\item{nEvents}{Number of events to occur in a simulated sequence of gradient change.}

\item{nSpecimens}{The number of specimens selected in each individual sample.}

\item{specimensPerTimestep}{The number of specimens returned in a given time-step by \code{getTimestepAbundances}, usually set to an unrealistically high number to represent the true 'unsampled' fossil assemblage. Default is 10000.}

\item{halfGradientOnly}{Whether to simulate only half of a background-event sequence, either beginning or terminating the simulation at the peak value. Only a single event can be simulated, so \code{nEvents} must be 1.}

\item{maxSampleTimeStep}{The}

\item{minSampleTimeStep}{The}

\item{includeInitialBackgroundPhase}{A logical indicating whether to include a lengthy background phase, for use in calibrating a simulation. This function is mainly for diagnostic purposes and may be removed in future updates.}

\item{plot}{Should the simulated time-series of fossil assemblages be shown as a sequence of generating and recovered gradient values against time? Default is \code{FALSE}.}

\item{thinOutput}{Should the output be thinned to just the sample properties and intrinsic variables? Default is FALSE.}
}
\value{
Returns a list, which by default has seven components: 
\code{implicitParameters}, the full list of parameters used for generating the simulated data; 
\code{simGradientChangeOut}, the simulated time-series of gradient change output by \code{setupSimulatedGradientChange};
\code{maxTime}, the total duration of the entire simulated time-series from start to end;
\code{simTimeVar}, a data frame specifying time-steps, sedimentary depth and environmental gradient values for simulating a time-series of sampled fossil assemblages, used as input in \code{\link{sampleFossilSeries}};
\code{fossilSeries}, a list containing the simulated time-series of sampled fossil assemblages from \code{\link{sampleFossilSeries}},
\code{ecology}, the recovered ecological variables for each simulated sample, as returned by \code{quantifyCommunityEcology},
and \code{sampleProperties}, a list containing a number of variables specific to individual .

If \code{thinList = TRUE} is used, then the output list
 contains only two components: 
 \code{sampleProperties} and \code{implicitParameters}.
 The \code{implicitParameters} component is the same as in the full output,
 but the \code{sampleProperties} component only contains information on when 
 (in both time and sedimentary depth) a given sample is located in the 
 simulated time-series, and the variable \code{scoreDCA1_recovered}.
}
\description{
Given a set of parameters and models describing species abundance,
stochastically models changes in an underlying biotic gradient and simulates
ecological change and a sequence of samples representing change in recovered
fossil assemblages over that interval, 
including estimating the recovered gradient via \code{quantifyCommunityEcology}.
}
\details{
Different parameterizations may be given as input, allowing different parameters to be unspecified.
Missing paramters are then calculated from the specified ones using \code{\link{calculateImplicitParameters}}.
}
\seealso{
\code{\link{calculateImplicitParameters}}
}
