% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeSimulationDCA.R
\name{summarizeSimulationDCA}
\alias{summarizeSimulationDCA}
\title{Summarize Simulation DCA}
\usage{
summarizeSimulationDCA(
  sampleData,
  nEvents,
  backgroundUpperBound,
  eventMagnitudesAbsolute,
  eventLowerBound,
  eventUpperBound
)
}
\arguments{
\item{sampleData}{}

\item{nEvents}{}

\item{backgroundUpperBound}{}

\item{eventMagnitudesAbsolute}{}

\item{eventLowerBound}{}

\item{eventUpperBound}{}
}
\description{
Summarize Simulation DCA
}
